+++ vimgolf/commands/__init__.py
@@ -1,0 +1,5 @@
+from .config import config
+from .local import local
+from .ls import list_
+from .put import put
+from .show import show
